refactor: 이벤트 등록 함수, 렌더 관련 함수 전부 모듈화
